refact: move components and screens to presentation
