+++ app/src/main/java/com/github/felipecastilhos/pokedexandroid/features/pokemon/presentation/components/PokemonTypeBadge.kt
@@ -1,4 +1,4 @@
-package com.github.felipecastilhos.pokedexandroid.features.pokemon.ui.components
+package com.github.felipecastilhos.pokedexandroid.features.pokemon.presentation.components
 
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
