+++ app/src/main/java/com/github/felipecastilhos/pokedexandroid/features/pokemon/presentation/viewmodel/PokemonDetailViewModel.kt
@@ -1,4 +1,4 @@
-package com.github.felipecastilhos.pokedexandroid.features.pokemon.domain.viewmodel
+package com.github.felipecastilhos.pokedexandroid.features.pokemon.presentation.viewmodel
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
