+++ app/src/main/java/com/github/felipecastilhos/pokedexandroid/features/pokemon/presentation/viewmodel/PokemonListViewModel.kt
@@ -1,4 +1,4 @@
-package com.github.felipecastilhos.pokedexandroid.features.pokemon.domain.viewmodel
+package com.github.felipecastilhos.pokedexandroid.features.pokemon.presentation.viewmodel
 
 import androidx.annotation.DrawableRes
 import androidx.lifecycle.ViewModel
